[frontend] moved value view to separate folder
