+++ frontend/src/view/Value.vue
